Separate all classes in Entities js into individual files
